# Documents/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "COSC2640 A3 API",
    "description": "The documentation and testing tool for COSC2640 A3 Microservices API.",
    "version": "v1.0"
  },
  "paths": {
    "/account/update-student": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "For student. Update student data including school name, faculty, and personal URL.",
        "description": "Request signature:\r\n\r\n<pre>\rPUT /account/update-student\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\rBody\r    {\r        \"id\": string,\r        \"schoolName\": string,\r        \"faculty\": string\r        \"personalUrl\": string\r    }\r</pre>\r\n",
        "parameters": [
          {
            "name": "accountId",
            "in": "header",
            "description": "The account's ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The required data to update Student.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDetail"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDetail"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDetail"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/account/student": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "For student. Get all student details including student data and account data.",
        "description": "Request signature:\r\n\r\n<pre>\rGET /account/student\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\r</pre>\r\n\r\n\r\nReturned object signature:\r\n\r\n<pre>\r{\r    email: string,\r    username: string,\r    phoneNumber: string,\r    phoneNumberConfirmed: boolean,\r    twoFaEnabled: boolean,\r    preferredName: string,\r    twoFa: {\r        qrImageUrl: string,\r        manualQrCode: string\r    },\r    studentId: string,\r    schoolName: string,\r    faculty: string,\r    personalUrl: string\r}\r</pre>\r\n",
        "parameters": [
          {
            "name": "accountId",
            "in": "header",
            "description": "The account's ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/account/update-teacher": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "For teacher. Update teacher data including company, jobTitle, and personal website.",
        "description": "Request signature:\r\n\r\n<pre>\rPUT /account/update-teacher\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\rBody\r    {\r        \"id\": string,\r        \"company\": string,\r        \"jobTitle\": string\r        \"personalUrl\": string\r    }\r</pre>\r\n",
        "parameters": [
          {
            "name": "accountId",
            "in": "header",
            "description": "The account's ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The required data to update Teacher.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDetail"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDetail"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDetail"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/account/teacher": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "For teacher. Get all teacher details including teacher data and account data.",
        "description": "Request signature:\r\n\r\n<pre>\rGET /account/teacher\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\r</pre>\r\n\r\n\r\nReturned object signature:\r\n\r\n<pre>\r{\r    email: string,\r    username: string,\r    phoneNumber: string,\r    phoneNumberConfirmed: boolean,\r    twoFaEnabled: boolean,\r    preferredName: string,\r    twoFa: {\r        qrImageUrl: string,\r        manualQrCode: string\r    },\r    teacherId: string,\r    company: string,\r    jobTitle: string,\r    personalWebsite: string\r}\r</pre>\r\n",
        "parameters": [
          {
            "name": "accountId",
            "in": "header",
            "description": "The account's ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/account/new-tfa": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "For both. To turn on or renew Two-Factor Authentication protection.",
        "description": "Request signature:\r\n\r\n<pre>\rPUT /account/new-tfa\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\r</pre>\r\n\r\n\r\nReturned object signature:\r\n\r\n<pre>\r{\r    qrImageUrl: string,\r    manualEntryKey: string\r}\r</pre>\r\n",
        "parameters": [
          {
            "name": "accountId",
            "in": "header",
            "description": "The account's ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/account/disable-tfa/{recaptchaToken}": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "For both. To turn off Two-Factor Authentication protection.",
        "description": "Request signature:\r\n\r\n<pre>\rPUT /account/disable-tfa/{string}\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\r</pre>\r\n",
        "parameters": [
          {
            "name": "accountId",
            "in": "header",
            "description": "The account's ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recaptchaToken",
            "in": "path",
            "description": "The recaptcha confirmation, not required in testings.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/account/set-phone-number/{phoneNumber}": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "For both. To add a phone number for an account. An SMS will be sent to the phone number to confirm it so use a <b>REAL</b> phone number.",
        "description": "Request signature:\r\n\r\n<pre>\rPUT /account/set-phone-number/{string}\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\r</pre>\r\n",
        "parameters": [
          {
            "name": "accountId",
            "in": "header",
            "description": "The account's ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phoneNumber",
            "in": "path",
            "description": "The phone number to be added.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
    "/account/remove-phone-number": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "For both. To remove phone number for an account.",
        "description": "Request signature:\r\n\r\n<pre>\rPUT /account/remove-phone-number\rHeaders\r    \"AccountId\": string\r    \"Authorization\": \"Bearer token\"\r</pre>\r\n",
        "parameters": [
          {
            "name": "accountId",
            "in": "header",
            "description": "The account's ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successfully processed."
          },
          "401": {
            "description": "Authorization failed: expired or mismatched or insufficient."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "workplace": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "personalUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}